requirements.txt
appdirs==1.4.4
black==21.6b0
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
distlib==0.3.2
filelock==3.0.12
flake8==3.9.2
identify==2.2.10
idna==2.10
isort==5.9.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.8.1
pre-commit==2.13.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.18.0
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.4.7
